requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
decouple==0.0.7
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docker==7.1.0
drf-yasg==1.21.9
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
requests==2.31.0
setuptools==75.8.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
